+++ app/templates/src/app/_config.js
@@ -1,0 +1,6 @@
+'use strict';
+
+angular
+  .module('<%= appname %>.common')
+  .constant('AppConfig', <%= APP_CONFIG %>)
+;