+++ app/templates/tasks/watch.js
@@ -1,12 +1,16 @@
 'use strict';
 
 var gulp = require('gulp');
+var browserSync = require('browser-sync');
 
 var args = require('./utils/cli-args');
 var paths = require('./utils/paths');
 var string = require('./utils/string');
 
-var entries = paths.sources.entries;
+var reload = browserSync.reload;
+var dist = paths.dist;
+var sources = paths.sources;
+var entries = sources.entries;
 var env = args.env || 'dev';
 var config = string.compile(paths.sources.config, {
   env: env
@@ -14,7 +18,7 @@ var config = string.compile(paths.sources.config, {
 
 module.exports = function() {
   // Watch all application scripts
-  gulp.watch([paths.sources.scripts], [
+  gulp.watch([sources.scripts], [
     'lint',
     'scripts.app'
   ]);
@@ -23,23 +27,30 @@ module.exports = function() {
     'scripts.vendor'
   ]);
   // Watch styles
-  gulp.watch([paths.sources.styles], [
+  gulp.watch([sources.styles], [
     'styles'
   ]);
   // Watch html files
   gulp.watch([entries.html], [
     'html'
   ]);
-  // Watch templates
-  gulp.watch([paths.sources.partials], [
+  // Watch views
+  gulp.watch([sources.partials], [
     'templates'
   ]);
   // Watch images
-  gulp.watch([paths.sources.images], [
+  gulp.watch([sources.images], [
     'images'
   ]);
   // Watch config
   gulp.watch([config], [
     'config'
   ]);
+  // Watch for live reload
+  gulp.watch([
+    dist.root + '*.html',
+    dist.scripts + '*.js',
+    dist.images + '*.*',
+    dist.fonts + '*.*'
+  ]).on('change', reload);
 };